// src/app/core/models/ensured.models.ts
export interface Ensured {
    dni: number;
    id_history: string;
    name: string;
    lastname_p: string;
    lastname_m: string;
    birthday: Date;
    gender: string;
    phone: number;
    email:string;
    civil_status: string;
    country: string;
    department: string;   
    district: string;
    address: string; 
  }